Enhancement: Changed how store names are assigned.
Previous: If no store name is provided in request, default store name is used. If no default store name is provided, an error is generated.  Store name is mandatory.

Current: If no store name is provided in request, default store name is used. If not default store name is provided, layer name is used for store name. Store name is optional.